Enterprise Intelligent Assistant/requirements.txt
streamlit
transformers
torch
sentence-transformers
scikit-learn
nltk
PyPDF2
python-docx
